extract_cron_job/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
fake-useragent==1.5.1
idna==3.7
numpy==1.26.4
pandas==2.2.2
pendulum==3.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
time-machine==2.14.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
